requirements.txt
AaioAsync==0.1.9
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
motor==3.4.0
multidict==6.0.5
pillow==10.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.3
python-dotenv==1.0.1
typing_extensions==4.11.0
yarl==1.9.4
